third_party/WebKit/LayoutTests/media/track/track-kind-user-preference-expected.txt

Test that user preference for text track kind is honored
RUN(internals.setUserPreferredLanguages(['jp', 'es', 'en', 'fr']))
EVENT(canplaythrough)

EXPECTED (tracks.length == '7') OK

EXPECTED (expectedTrack.mode == 'showing') OK
EXPECTED (expectedTrack.kind == 'subtitles') OK
EXPECTED (expectedTrack.language == 'ar') OK

EXPECTED (expectedTrack.mode == 'showing') OK
EXPECTED (expectedTrack.kind == 'subtitles') OK
EXPECTED (expectedTrack.language == 'es') OK

EXPECTED (expectedTrack.mode == 'showing') OK
EXPECTED (expectedTrack.kind == 'captions') OK
EXPECTED (expectedTrack.language == 'es') OK

EVENT(load)
EXPECTED (track.readyState == '2') OK
EXPECTED (expectedTrack.mode == 'showing') OK
EXPECTED (expectedTrack.kind == 'subtitles') OK
EXPECTED (expectedTrack.language == 'jp') OK

END OF TEST
